Added placement request test repository
Moved some test-only methods from the real repo to the test repo
